+++ lua/plugins/lsp.lua
@@ -7,8 +7,9 @@ return {
     "neovim/nvim-lspconfig",
     event = "BufReadPre",
     dependencies = {
-      "mason.nvim",
-      "mason-lspconfig.nvim",
+      "williamboman/mason.nvim",
+      "williamboman/mason-lspconfig.nvim",
+      "hrsh7th/cmp-nvim-lsp",
     },
     config = function()
       require("lspconfig.ui.windows").default_options.border = "rounded"
@@ -19,7 +20,9 @@ return {
       lsp_defaults.capabilities =
           vim.tbl_deep_extend("force", lsp_defaults.capabilities, require("cmp_nvim_lsp").default_capabilities())
 
+      local lsp_group = vim.api.nvim_create_augroup("NikaVimLspAttach", { clear = true })
       vim.api.nvim_create_autocmd("LspAttach", {
+        group = lsp_group,
         callback = function(event)
           local opts = { buffer = event.buf, silent = true }
 
@@ -31,7 +34,14 @@ return {
           vim.keymap.set("n", "gr", vim.lsp.buf.references, opts)
           vim.keymap.set("n", "gs", vim.lsp.buf.signature_help, opts)
           vim.keymap.set("n", "<F2>", vim.lsp.buf.rename, opts)
-          vim.keymap.set({ "n", "x" }, "<F3>", vim.lsp.buf.format, opts)
+          vim.keymap.set({ "n", "x" }, "<F3>", function()
+            local ok, conform = pcall(require, "conform")
+            if ok then
+              conform.format({ async = true, lsp_format = "fallback" })
+            else
+              vim.lsp.buf.format({ async = true })
+            end
+          end, opts)
           vim.keymap.set("n", "<F4>", vim.lsp.buf.code_action, opts)
 
           vim.lsp.inlay_hint.enable(true, { bufnr = event.buf })
@@ -62,21 +72,13 @@ return {
   -- Mason LSP Config
   {
     "williamboman/mason-lspconfig.nvim",
-    dependencies = "mason.nvim",
+    dependencies = {
+      "williamboman/mason.nvim",
+      "neovim/nvim-lspconfig",
+    },
     config = function()
       require("mason-lspconfig").setup({
-        ensure_installed = {
-          "lua_ls",
-          "pyright",
-          "ts_ls",
-          "html",
-          "cssls",
-          "jsonls",
-          "yamlls",
-          "bashls",
-          "clangd",
-          "rust_analyzer",
-        },
+        automatic_installation = false,
         handlers = {
           function(server_name)
             require("lspconfig")[server_name].setup({})