+++ lua/plugins/project.lua
@@ -2,6 +2,7 @@ return {
   {
     "ahmedkhalf/project.nvim",
     event = "VeryLazy",
+    dependencies = { "nvim-telescope/telescope-project.nvim" },
     config = function()
       require("project_nvim").setup({
         detection_methods = { "pattern", "lsp" },