+++ lua/plugins/telescope.lua
@@ -4,7 +4,7 @@
 return {
   {
     "nvim-telescope/telescope.nvim",
-        branch = "0.1.x",
+    branch = "0.1.x",
     dependencies = {
       "nvim-lua/plenary.nvim",
       { "nvim-telescope/telescope-fzf-native.nvim", build = "make" },
@@ -46,7 +46,6 @@ return {
         },
       })
 
-      telescope.load_extension("fzf")
     end,
   },
 }